+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Chg #105: Change PHP constraint in `composer.json` to `8.1 - 8.5` (@rustamwin)
 - Enh #108: Bump minimal `yiisoft/html` version to `3.13` and add support for `^4.0` (@vjik)
 - Bug #113: Fix `array_merge()` argument order in `Menu::renderItem()` so that item-level `linkAttributes` override widget-level ones (@WarLikeLaux)
+- New #117: Add container wrapper and JSON-LD structured data to `Breadcrumbs` widget (@WarLikeLaux)
 
 ## 2.1.1 September 23, 2025